- deleteRollback and trimRollbackFile moved to the createRollbackFile. - createRollbackFile renamed to rollbackUtils. - Added test for parseRestoreArg function.
