+++ src/utils/utils.ts
@@ -1,11 +1,7 @@
 import { existsSync } from "fs";
 import {
   lstat,
-  readdir,
-  readFile,
-  rename,
-  unlink,
-  writeFile
+  readdir, rename
 } from "fs/promises";
 import { join, parse, resolve } from "path";
 import readline from "readline";
@@ -32,12 +28,8 @@ import type {
   NumberOfDuplicatedNames,
   PromiseRejectedWriteResult,
   RenameItem,
-  RenameItemsArray,
-  RollbackFile,
-  TrimRollbackFile,
-  TruncateFileName
+  RenameItemsArray, TruncateFileName
 } from "../types.js";
-import { checkRestoreFile } from "./restoreUtils.js";
 
 const {
   allRenameFailed,
@@ -48,7 +40,7 @@ const {
   noChildEntries,
 } = ERRORS.utils;
 const { truncateInvalidArgument } = ERRORS.transforms;
-const { noRollbackFile } = ERRORS.cleanRollback;
+
 const { failReport, failItem } = STATUS.settledPromisesEval;
 
 export const jsonParseReplicate = <T>(arg: string): T => JSON.parse(arg) as T;
@@ -113,91 +105,6 @@ export const extractCurrentReferences = (
     );
   }
   return { withIds: namesAndReferences, noIds: namesLeftToCheck };
-};
-
-export const trimRollbackFile: TrimRollbackFile = async ({
-  sourcePath,
-  targetLevel,
-  failed,
-}) => {
-  const targetDir = determineDir(sourcePath);
-  const targetPath = resolve(targetDir, ROLLBACK_FILE_NAME);
-  const rollBackFileExists = existsSync(targetPath);
-  if (!rollBackFileExists) {
-    throw new Error(noRollbackFile);
-  }
-  const rollbackFile = JSON.parse(await readFile(targetPath, "utf-8"));
-
-  const verifiedRollback = checkRestoreFile(rollbackFile);
-  const remainingTransforms = verifiedRollback.transforms.slice(targetLevel);
-
-  const shouldDelete = [remainingTransforms, failed].every(
-    (arr) => !arr.length
-  );
-
-  if (shouldDelete) {
-    process.stdout.write("Deleting rollback file...");
-    await unlink(targetPath);
-    process.stdout.write("DONE!");
-    return;
-  }
-
-  let mappedFailed: RenameItemsArray = [];
-  if (failed.length) {
-    const mappedEntry = new Map() as Map<string, RenameItem>;
-
-    remainingTransforms[0]?.reduce(
-      (map, curr) => map.set(curr.referenceId, curr),
-      mappedEntry
-    );
-
-    !mappedEntry.size ? mappedFailed = [...failed] :
-      failed.forEach(({ rename, original, referenceId }) => {
-        const ref = mappedEntry.get(referenceId);
-        if (!ref) return mappedFailed.push({ rename, original, referenceId });
-
-        const isDistinct = original !== ref.rename;
-          mappedFailed.push({
-            rename,
-            original: isDistinct ? ref.rename : original,
-            referenceId,
-          });
-      });
-    
-    console.log(
-      "Failed restore items will be appended to most recent rollback entry."
-    );
-  }
-
-  const newRollbackFile: RollbackFile = {
-    sourcePath,
-    transforms: [[...mappedFailed], ...remainingTransforms].filter(
-      (entry) => entry.length
-    ),
-  };
-
-  process.stdout.write("Updating rollback file...");
-  await writeFile(
-    resolve(targetDir, ROLLBACK_FILE_NAME),
-    JSON.stringify(newRollbackFile, undefined, 2),
-    "utf-8"
-  );
-
-  process.stdout.write("DONE!");
-};
-
-export const deleteRollbackFile = async (
-  transformPath?: string
-): Promise<void> => {
-  const targetDir = determineDir(transformPath);
-  const targetPath = resolve(targetDir, ROLLBACK_FILE_NAME);
-  const rollBackFileExists = existsSync(targetPath);
-  if (!rollBackFileExists) {
-    throw new Error(noRollbackFile);
-  }
-  process.stdout.write("Deleting rollback file...");
-  await unlink(targetPath);
-  process.stdout.write("DONE!");
 };
 
 /**Will separate the basename and file extension, in addition to providing
